fixes and changes for better structure and readability.

+++ NaiveBayes/src/Structure/Email.cs
@@ -6,7 +6,7 @@ public class Email : IEmail
     [Ignore]
     public String[]? TokenizedMessage;
 
-    public int Id { get; set; }
+    public double Id { get; set; }
     public string? Subject { get; set; }
     public string? Message { get; set; }
     [EnumIgnoreCase]
@@ -15,6 +15,7 @@ public class Email : IEmail
 
     public void PerformDeserializationLogic()
     {
+
         TokenizedMessage = TextProcessing.Tokenize(Message);
         TokenizedMessage = TextProcessing.RemoveStopWords(TokenizedMessage);
         TokenizedMessage = TextProcessing.RemovePunctuation(TokenizedMessage);
